Isolating stacks in privdata example
In the protecting private metadata example, there is a key which needs
to be kept secret for encrypting messages. The capability which contains
this key is a global sealed object. During the execution of the encrypt
function, the unsealed form of this capability can be spilled to the
stack, then found by reading stack pages from outside the function. To
prevent this, when calling the encrypt function, the stack pointer is
swapped for another pointer to memory which serves as the encryption
function's stack. When returning to caller code, the original stack
pointer is restored.

Change-Id: I59d2c0648d48c0be0437878ad06de3d3f82046ad